pyproject.toml
[project]
name = "segmenter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "earthengine-api>=1.7.4",
    "folium>=0.18.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pillow>=11.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.41.0",
    "streamlit-folium>=0.23.0",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
